requirements.txt
uvicorn==0.15.0
slackclient==2.9.3
python-dotenv==0.19.2
pydantic==1.9.0
fastapi==0.70.1
google-cloud-pubsub==2.9.0
